Split and improve the unit tests
