+++ plugins/woocommerce/tests/php/src/Internal/Admin/Settings/Mocks/FakePaymentGateway.php
@@ -50,11 +50,84 @@ class FakePaymentGateway extends \WC_Payment_Gateway {
 	public string $plugin_slug = 'fake-plugin-slug';
 
 	/**
+	 * Corresponding gateway plugin file.
+	 *
+	 * Skip the .php extension to match the format used by the WP API.
+	 *
+	 * @var string
+	 */
+	public string $plugin_file = 'fake-plugin-slug/fake-plugin-file';
+
+	/**
 	 * The recommended payment methods list.
 	 *
 	 * @var array
 	 */
 	public array $recommended_payment_methods = array();
+
+	/**
+	 * Whether or not this gateway still requires setup to function.
+	 *
+	 * @var bool
+	 */
+	public bool $needs_setup = false;
+
+	/**
+	 * The test mode.
+	 *
+	 * @var bool
+	 */
+	public bool $test_mode = false;
+
+	/**
+	 * The dev mode.
+	 *
+	 * @var bool
+	 */
+	public bool $dev_mode = false;
+
+	/**
+	 * Constructor.
+	 *
+	 * @param string $id Optional. The gateway ID.
+	 */
+	public function __construct( string $id = '', array $props = array() ) {
+		if ( ! empty( $id ) ) {
+			$this->id = $id;
+		}
+
+		// Go through the props and set them on the object.
+		foreach ( $props as $prop => $value ) {
+			$this->$prop = $value;
+		}
+	}
+
+	/**
+	 * Return whether or not this gateway still requires setup to function.
+	 *
+	 * @return bool
+	 */
+	public function needs_setup() {
+		return $this->needs_setup;
+	}
+
+	/**
+	 * Get the gateway settings page URL.
+	 *
+	 * @return string The gateway settings page URL.
+	 */
+	public function get_settings_url(): string {
+		return 'https://example.com/wp-admin/admin.php?page=wc-settings&tab=checkout&section=bogus_settings';
+	}
+
+	/**
+	 * Get the gateway onboarding start/continue URL.
+	 *
+	 * @return string The gateway onboarding start/continue URL.
+	 */
+	public function get_connection_url(): string {
+		return 'https://example.com/connection-url';
+	}
 
 	/**
 	 * Get the recommended payment methods list.
@@ -65,5 +138,23 @@ class FakePaymentGateway extends \WC_Payment_Gateway {
 	 */
 	public function get_recommended_payment_methods( string $country_code = '' ): array {
 		return $this->recommended_payment_methods;
+	}
+
+	/**
+	 * Check if the gateway is in test mode.
+	 *
+	 * @return bool True if the gateway is in test mode, false otherwise.
+	 */
+	public function is_test_mode(): bool {
+		return $this->test_mode;
+	}
+
+	/**
+	 * Check if the gateway is in dev mode.
+	 *
+	 * @return bool True if the gateway is in dev mode, false otherwise.
+	 */
+	public function is_dev_mode(): bool {
+		return $this->dev_mode;
 	}
 }
