+++ plugins/woocommerce/tests/php/src/Internal/Admin/Settings/PaymentsRestControllerTest.php
@@ -3,6 +3,7 @@ declare( strict_types=1 );
 
 namespace Automattic\WooCommerce\Tests\Internal\Admin\Settings;
 
+use Automattic\WooCommerce\Internal\Admin\Settings\PaymentProviders;
 use Automattic\WooCommerce\Internal\Admin\Settings\Payments;
 use Automattic\WooCommerce\Internal\Admin\Settings\PaymentsRestController;
 use PHPUnit\Framework\MockObject\MockObject;
@@ -132,12 +133,14 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 		$this->assertArrayHasKey( 'enabled', $provider['state'], 'Provider (gateway) `state[enabled]` entry is missing' );
 		$this->assertArrayHasKey( 'needs_setup', $provider['state'], 'Provider (gateway) `state[needs_setup]` entry is missing' );
 		$this->assertArrayHasKey( 'test_mode', $provider['state'], 'Provider (gateway) `state[test_mode]` entry is missing' );
-		$this->assertArrayHasKey( 'management', $provider, 'Provider (gateway) `management` entry is missing' );
-		$this->assertArrayHasKey( 'settings_url', $provider['management'], 'Provider (gateway) `management[settings_url]` entry is missing' );
+		$this->assertArrayHasKey( 'management', $provider, 'Gateway `management` entry is missing' );
+		$this->assertArrayHasKey( '_links', $provider['management'], 'Gateway `management[_links]` entry is missing' );
+		$this->assertArrayHasKey( 'settings', $provider['management']['_links'], 'Gateway `management[_links][settings]` entry is missing' );
 		$this->assertArrayHasKey( 'links', $provider, 'Provider (gateway) `links` entry is missing' );
 		$this->assertCount( 1, $provider['links'] );
 		$this->assertArrayHasKey( 'plugin', $provider, 'Provider (gateway) `plugin` entry is missing' );
 		$this->assertArrayHasKey( 'slug', $provider['plugin'], 'Provider (gateway) `plugin[slug]` entry is missing' );
+		$this->assertArrayHasKey( 'file', $provider['plugin'], 'Provider (gateway) `plugin[file]` entry is missing' );
 		$this->assertArrayHasKey( 'status', $provider['plugin'], 'Provider (gateway) `plugin[status]` entry is missing' );
 		$this->assertArrayHasKey( '_links', $provider, 'Provider (gateway) `_links` entry is missing' );
 
@@ -212,7 +215,7 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 			array(
 				'_wc_pes_woopayments',
 				'_wc_pes_paypal_full_stack',
-				Payments::OFFLINE_METHODS_ORDERING_GROUP,
+				PaymentProviders::OFFLINE_METHODS_ORDERING_GROUP,
 				'paypal',
 			),
 			array_column( $data['providers'], 'id' )
@@ -286,7 +289,7 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 			array(
 				'_wc_pes_woopayments',
 				'_wc_pes_paypal_full_stack',
-				Payments::OFFLINE_METHODS_ORDERING_GROUP,
+				PaymentProviders::OFFLINE_METHODS_ORDERING_GROUP,
 				'paypal',
 			),
 			array_column( $data['providers'], 'id' )
@@ -382,7 +385,7 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 		// Assert that we have the right providers, in the right order.
 		$this->assertSame(
 			array(
-				Payments::OFFLINE_METHODS_ORDERING_GROUP,
+				PaymentProviders::OFFLINE_METHODS_ORDERING_GROUP,
 				'paypal',
 			),
 			array_column( $data['providers'], 'id' )
@@ -841,7 +844,7 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 			$mock_providers[] = array(
 				'id'                => '_wc_pes_woopayments',
 				'_order'            => $order++,
-				'_type'             => Payments::PROVIDER_TYPE_SUGGESTION,
+				'_type'             => PaymentProviders::TYPE_SUGGESTION,
 				'title'             => 'Accept payments with Woo',
 				'description'       => 'With WooPayments, you can securely accept major cards, Apple Pay, and payments in over 100 currencies. Track cash flow and manage recurring revenue directly from your store’s dashboard - with no setup costs or monthly fees.',
 				'short_description' => 'Credit/debit cards, Apple Pay, Google Pay and more.',
@@ -888,7 +891,7 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 			$mock_providers[] = array(
 				'id'                => '_wc_pes_paypal_full_stack',
 				'_order'            => $order++,
-				'_type'             => Payments::PROVIDER_TYPE_SUGGESTION,
+				'_type'             => PaymentProviders::TYPE_SUGGESTION,
 				'title'             => 'PayPal Payments',
 				'description'       => 'Safe and secure payments using credit cards or your customer&#039;s PayPal account.',
 				'short_description' => '',
@@ -929,7 +932,7 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 			$mock_providers[] = array(
 				'id'          => '_wc_offline_payment_methods_group',
 				'_order'      => $order++,
-				'_type'       => Payments::PROVIDER_TYPE_OFFLINE_PMS_GROUP,
+				'_type'       => PaymentProviders::TYPE_OFFLINE_PMS_GROUP,
 				'title'       => 'Offline Payment Methods',
 				'description' => 'Allow shoppers to pay offline.',
 				'icon'        => 'http://localhost:8888/wp-content/plugins/woocommerce/assets/images/payment_methods/cod.svg',
@@ -937,14 +940,16 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 			$mock_providers[] = array(
 				'id'          => 'bacs',
 				'_order'      => $order++,
-				'_type'       => Payments::PROVIDER_TYPE_OFFLINE_PM,
+				'_type'       => PaymentProviders::TYPE_OFFLINE_PM,
 				'title'       => 'Direct bank transfer',
 				'description' => 'Take payments in person via BACS. More commonly known as direct bank/wire transfer.',
 				'supports'    => array(
 					'products',
 				),
 				'plugin'      => array(
+					'_type'  => 'wporg',
 					'slug'   => 'woocommerce',
+					'file'   => 'woocommerce/woocommerce',
 					'status' => 'active',
 				),
 				'icon'        => 'http://localhost:8888/wp-content/plugins/woocommerce/assets/images/payment_methods/bacs.svg',
@@ -961,20 +966,26 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 					'dev_mode'    => false,
 				),
 				'management'  => array(
-					'settings_url' => 'http://localhost:8888/wp-admin/admin.php?page=wc-settings&tab=checkout&section=bacs',
+					'_links' => array(
+						'settings' => array(
+							'href' => 'http://localhost:8888/wp-admin/admin.php?page=wc-settings&tab=checkout&section=bacs',
+						),
+					),
 				),
 			);
 			$mock_providers[] = array(
 				'id'          => 'cheque',
 				'_order'      => $order++,
-				'_type'       => Payments::PROVIDER_TYPE_OFFLINE_PM,
+				'_type'       => PaymentProviders::TYPE_OFFLINE_PM,
 				'title'       => 'Check payments',
 				'description' => 'Take payments in person via checks. This offline gateway can also be useful to test purchases.',
 				'supports'    => array(
 					'products',
 				),
 				'plugin'      => array(
+					'_type'  => 'wporg',
 					'slug'   => 'woocommerce',
+					'file'   => 'woocommerce/woocommerce',
 					'status' => 'active',
 				),
 				'icon'        => 'http://localhost:8888/wp-content/plugins/woocommerce/assets/images/payment_methods/cheque.svg',
@@ -991,20 +1002,26 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 					'dev_mode'    => false,
 				),
 				'management'  => array(
-					'settings_url' => 'http://localhost:8888/wp-admin/admin.php?page=wc-settings&tab=checkout&section=cheque',
+					'_links' => array(
+						'settings' => array(
+							'href' => 'http://localhost:8888/wp-admin/admin.php?page=wc-settings&tab=checkout&section=cheque',
+						),
+					),
 				),
 			);
 			$mock_providers[] = array(
 				'id'          => 'cod',
 				'_order'      => $order++,
-				'_type'       => Payments::PROVIDER_TYPE_OFFLINE_PM,
+				'_type'       => PaymentProviders::TYPE_OFFLINE_PM,
 				'title'       => 'Cash on delivery',
 				'description' => 'Let your shoppers pay upon delivery — by cash or other methods of payment.',
 				'supports'    => array(
 					'products',
 				),
 				'plugin'      => array(
+					'_type'  => 'wporg',
 					'slug'   => 'woocommerce',
+					'file'   => 'woocommerce/woocommerce',
 					'status' => 'active',
 				),
 				'icon'        => 'http://localhost:8888/wp-content/plugins/woocommerce/assets/images/payment_methods/cod.svg',
@@ -1021,7 +1038,11 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 					'dev_mode'    => false,
 				),
 				'management'  => array(
-					'settings_url' => 'http://localhost:8888/wp-admin/admin.php?page=wc-settings&tab=checkout&section=cod',
+					'_links' => array(
+						'settings' => array(
+							'href' => 'http://localhost:8888/wp-admin/admin.php?page=wc-settings&tab=checkout&section=cod',
+						),
+					),
 				),
 			);
 		}
@@ -1030,7 +1051,7 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 			$mock_providers[] = array(
 				'id'          => 'paypal',
 				'_order'      => $order++,
-				'_type'       => Payments::PROVIDER_TYPE_GATEWAY,
+				'_type'       => PaymentProviders::TYPE_GATEWAY,
 				'title'       => 'PayPal',
 				'description' => 'PayPal',
 				'supports'    => array( 'products' ),
@@ -1041,7 +1062,11 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 					'dev_mode'    => false,
 				),
 				'management'  => array(
-					'settings_url' => 'admin.php?page=wc-settings&tab=checkout&section=paypal',
+					'_links' => array(
+						'settings' => array(
+							'href' => 'admin.php?page=wc-settings&tab=checkout&section=paypal',
+						),
+					),
 				),
 				'image'       => 'https://example.com/image.png',
 				'icon'        => 'https://example.com/icon.png',
@@ -1054,6 +1079,7 @@ class PaymentsRestControllerTest extends WC_REST_Unit_Test_Case {
 				'plugin'      => array(
 					'_type'  => 'wporg',
 					'slug'   => 'woocommerce',
+					'file'   => 'woocommerce/woocommerce',
 					'status' => 'active',
 				),
 			);
